# Fix case issues with Params