@bep 2016-11-22
There seems to be a slight performance penalty with this:

```
benchmark           old ns/op      new ns/op       delta
BenchmarkHugo-4     9836140771     10273595307     +4.45%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     41718655       42182280       +1.11%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     8887030688     8895689976     +0.10%
```

Will have to look into this a little.