@bep 2016-11-22
OK, I have removed a obvious ineffeciency, and this may be livable:

```
benchmark           old ns/op      new ns/op       delta
BenchmarkHugo-4     9908842113     10242229016     +3.36%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     41717487       42137686       +1.01%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     8865691320     8841285632     -0.28%
```