@bep 2016-11-22
Now it is "fast enough":

```
benchmark           old ns/op      new ns/op      delta
BenchmarkHugo-4     9795843285     9700004476     -0.98%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     41717550       42134335       +1.00%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     8870693120     8899367200     +0.32%
´``