@spiegel-im-spiegel 2016-11-22
2016年11月22日(火) 18:08 Bjørn Erik Pedersen <notifications@github.com>:

> This replaces #2630 <https://github.com/spf13/hugo/pull/2630> -- but it
> needs some more testing before we try again. There is a inifinite loop
> lurking in here ...
>
> There are currently several Params and case related issues floating around
> in Hugo.
>
> This is very confusing for users and one of the most common support
> questions on the forum.
>
> And while there have been done some great leg work in Viper etc., this is
> of limited value since this and similar doesn't work:
>
> Params.myCamelCasedParam
>
> Hugo has control over all the template method invocations, and can take
> care of all the lower-casing of the map lookup keys.
>
> But that doesn't help with direct template lookups of type
> Site.Params.TWITTER_CONFIG.USER_ID.
>
> This commit solves that by doing some carefully crafted modifications of
> the templates' AST -- lowercasing the params keys.
>
> This is low-level work, but it's not like the template API wil change --
> and this is important enough to defend such "bit fiddling".
>
> Tests are added for all the template engines: Go templates, Ace and Amber.
>
> Fixes #2615 <https://github.com/spf13/hugo/issues/2615>
> Fixes #1129 <https://github.com/spf13/hugo/issues/1129>
> Fixes #2590 <https://github.com/spf13/hugo/issues/2590>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/spf13/hugo/pull/2717
> Commit Summary
>
>    - Fix case issues with Params
>
> File Changes
>
>    - *A* hugolib/case_insensitive_test.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-0> (289)
>    - *M* hugolib/page.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-1> (3)
>    - *M* hugolib/site.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-2> (2)
>    - *M* hugolib/site_render.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-3> (2)
>    - *M* tpl/template.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-4> (29)
>    - *A* tpl/template_ast_transformers.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-5> (227)
>    - *A* tpl/template_ast_transformers_test.go
>    <https://github.com/spf13/hugo/pull/2717/files#diff-6> (218)
>
> Patch Links:
>
>    - https://github.com/spf13/hugo/pull/2717.patch
>    - https://github.com/spf13/hugo/pull/2717.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/spf13/hugo/pull/2717>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AHNbtUUY1vmzcXaNxackuvoPy8RlghsIks5rArEBgaJpZM4K5KRS>
> .
>
-- 

-- 
Ysuhiro ARAKAWA <spiegel.im.spiegel@gmail.com>